google-analytics-mcp/streamlit_ui/requirements.txt
streamlit
google-analytics-data
openai
pandas 